+++ src/utility.jl
@@ -37,10 +37,10 @@ function init_system(
     )
     arr = arrowhead_matrix(Δ)
     n_sites = size(arr, 1)
-    H_nat, n_occ = to_natural_orbitals(arr)
+    H_nat, n_occ = natural_impurity_orbital(arr)
     n_bit, V_v, V_c = get_RAS_parameters(n_sites, n_occ, L_c, L_v)
     fs = FockSpace(Orbitals(n_bit), FermionicSpin(1 // 2))
-    H = natural_orbital_ras_operator(H_nat, H_int, ϵ_imp, fs, n_occ, L_v, L_c, p)
+    H = natural_impurity_orbital_ras_operator(H_nat, H_int, ϵ_imp, fs, n_occ, L_v, L_c, p)
     ψ_start = RASWavefunction_singlet(Dict{UInt64, Float64}, L_v, L_c, V_v, V_c, p)
     E0, ψ0 = ground_state!(H, ψ_start, 5, typemax(Int), var)
     return H, E0, ψ0