+++ test/wavefunctions.jl
@@ -56,14 +56,14 @@ using Test
         n_sites = 1 + n_bath
 
         Δ = hybridization_function_bethe_simple(n_bath)
-        H_nat, n_occ = to_natural_orbitals(arrowhead_matrix(Δ))
+        H_nat, n_occ = natural_impurity_orbital(arrowhead_matrix(Δ))
         n_bit, V_v, V_c = get_RAS_parameters(n_sites, n_occ, L_c, L_v)
 
         # RASWavefunction
         fs = FockSpace(Orbitals(n_bit), FermionicSpin(1 // 2))
         n = occupations(fs)
         H_int = U * n[1, -1 // 2] * n[1, 1 // 2]
-        H = natural_orbital_ras_operator(H_nat, H_int, -μ, fs, n_occ, L_v, L_c, p)
+        H = natural_impurity_orbital_ras_operator(H_nat, H_int, -μ, fs, n_occ, L_v, L_c, p)
         ψ_start = RASWavefunction_singlet(Dict{UInt64, Float64}, L_v, L_c, V_v, V_c, p)
 
         # 10 steps total
@@ -91,7 +91,7 @@ using Test
         fs = FockSpace(Orbitals(n_sites), FermionicSpin(1 // 2))
         n = occupations(fs)
         H_int = U * n[1, -1 // 2] * n[1, 1 // 2]
-        H_wf = natural_orbital_operator(H_nat, H_int, -μ, fs, n_occ, L_v, L_c)
+        H_wf = natural_impurity_orbital_operator(H_nat, H_int, -μ, fs, n_occ, L_v, L_c)
         Fermions.shift_spectrum!(H_wf, E0)
         ψ0_wf = Wavefunction(ψ0)
         foo = H_wf * ψ0_wf